moved eventListHome from shared to widgets
